% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{batchValidateCoy}
\alias{batchValidateCoy}
\title{Returns a vector of booleans for each coy item in the input}
\usage{
batchValidateCoy(x, checkComma = TRUE)
}
\arguments{
\item{x}{list of coy strings}

\item{checkComma}{flag for toggling validation of coy starting character}
}
\value{
vector of booleans
}
\description{
Returns a vector of booleans for each coy item in the input
}
