% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{budBreak}
\alias{budBreak}
\title{Return the percentage of buds that have shot (floral or vegetative)}
\usage{
budBreak(x)
}
\arguments{
\item{x}{A valid coy string}
}
\value{
a double
}
\description{
Return the percentage of buds that have shot (floral or vegetative)
}
