% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{calculateKFperWB}
\alias{calculateKFperWB}
\title{Calculate the average king flowers per winter bud metric}
\usage{
calculateKFperWB(x)
}
\arguments{
\item{x}{a valid coy string}
}
\value{
numeric vector
}
\description{
Calculate the average king flowers per winter bud metric
}
