% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{calculateLperWB}
\alias{calculateLperWB}
\title{Calculate average lateral flowers per winter bud}
\usage{
calculateLperWB(x)
}
\arguments{
\item{x}{a valid coy string}
}
\value{
a numeric vector
}
\description{
Calculate average lateral flowers per winter bud
}
