% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{checkChars}
\alias{checkChars}
\title{Check allowed character types}
\usage{
checkChars(x)
}
\arguments{
\item{x}{A valid coy string}
}
\description{
Check all allowed character types in components of yield string
 [,.0-9Ll] are all allowed characters
}
