% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{coyChars}
\alias{coyChars}
\title{Find malformed characters in a components of yield string}
\usage{
coyChars(x)
}
\arguments{
\item{x}{A valid coy string}
}
\description{
Uses checkChars(...) and returns a list
}
