% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{CoyProcessor}
\alias{CoyProcessor}
\title{Returns a data frame with the standard COY analysis}
\usage{
CoyProcessor(data, name = "coy")
}
\arguments{
\item{x}{a data frame containing coy strings and their ID}
}
\value{
a data frame
}
\description{
Returns a data frame with the standard COY analysis
}
