% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{coyStart}
\alias{coyStart}
\title{Is the start of a coy string valid?}
\usage{
coyStart(x)
}
\arguments{
\item{x}{A valid coy string}
}
\description{
Pattern match the expected start of a components of yield string
}
