% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{extractFloralShoots}
\alias{extractFloralShoots}
\title{Return a string with only floral shoots}
\usage{
extractFloralShoots(x)
}
\arguments{
\item{x}{a valid COY string}
}
\value{
a string with only numerical values 1-9
}
\description{
Return a string with only floral shoots
}
