% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{floralPercentage}
\alias{floralPercentage}
\title{Calculate the percentage of floral shoots, automatically excludes bird damage}
\usage{
floralPercentage(x)
}
\arguments{
\item{x}{a valid coy string}
}
\value{
numeric vector
}
\description{
Calculate the percentage of floral shoots, automatically excludes bird damage
}
