% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{FloralShoots}
\alias{FloralShoots}
\title{Extract FloralShoots component}
\usage{
FloralShoots(x)
}
\arguments{
\item{x}{A valid coy string}
}
\value{
numeric vector
}
\description{
Extract FloralShoots component
}
