% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{hasBirdDamage}
\alias{hasBirdDamage}
\title{Returns a boolean value corresponding to whether the coy string contains chaacters for bird damage}
\usage{
hasBirdDamage(x)
}
\arguments{
\item{x}{a valid coy string}
}
\value{
a boolean
}
\description{
Returns a boolean value corresponding to whether the coy string contains chaacters for bird damage
}
