% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{isValidCoy}
\alias{isValidCoy}
\title{Return a logical vector corresponding to the validity of input string}
\usage{
isValidCoy(x, checkComma = TRUE)
}
\arguments{
\item{x}{coy string}

\item{checkComma}{flag to toggle validation of the starting character}
}
\value{
a boolean vector
}
\description{
Return a logical vector corresponding to the validity of input string
}
