% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{KingFlowersPerWinterBud}
\alias{KingFlowersPerWinterBud}
\title{Calculate the average king flowers per winter bud metric}
\usage{
KingFlowersPerWinterBud(x)
}
\arguments{
\item{x}{a valid coy string}
}
\value{
numeric vector
}
\description{
Calculate the average king flowers per winter bud metric
}
