% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{LateralFlowers}
\alias{LateralFlowers}
\title{Extract LateralFlowers  component}
\usage{
LateralFlowers(x)
}
\arguments{
\item{x}{A valid coy string}
}
\value{
numeric vector
}
\description{
Extract LateralFlowers  component
}
