% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{okChars}
\alias{okChars}
\title{Test if characters are all ok}
\usage{
okChars(x)
}
\arguments{
\item{x}{A valid coy string}
}
\description{
Uses checkChars(...) and returns a logical if all characters are ok
}
