% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{removeBirdDamage}
\alias{removeBirdDamage}
\title{Return coy string wth all bird damage values removed}
\usage{
removeBirdDamage(x)
}
\arguments{
\item{x}{a valid coy string}
}
\value{
coy string with no bird damage
}
\description{
Return coy string wth all bird damage values removed
}
