% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{splitChars}
\alias{splitChars}
\title{Split components of yields characters into a string}
\usage{
splitChars(x)
}
\arguments{
\item{x}{A valid coy string}
}
\description{
Split components of yields characters into a string
}
