% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{validateCOY}
\alias{validateCOY}
\title{Checks and validates a component of yield string}
\usage{
validateCOY(x)
}
\arguments{
\item{x}{A valid coy string}
}
\description{
Checks and validates a component of yield string
}
