% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{vegetativePercentage}
\alias{vegetativePercentage}
\title{Calculate vegetative shoot percentage, automatically excludes bird damage}
\usage{
vegetativePercentage(x)
}
\arguments{
\item{x}{a vlid coy string}
}
\value{
numeric vector
}
\description{
Calculate vegetative shoot percentage, automatically excludes bird damage
}
