% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{VegetShoots}
\alias{VegetShoots}
\title{Count VegetShoots component}
\usage{
VegetShoots(x)
}
\arguments{
\item{x}{A valid coy string}
}
\value{
numeric vector
}
\description{
Count VegetShoots component
}
