% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{wideCOY}
\alias{wideCOY}
\title{Return a dataframe in wide format of components of yield}
\usage{
wideCOY(x, names = NULL)
}
\arguments{
\item{x}{A valid coy string}
}
\value{
numeric vector
}
\description{
Return a dataframe in wide format of components of yield
}
