% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coy.R
\name{CoyProcessor}
\alias{CoyProcessor}
\title{Returns a data frame with the standard COY analysis}
\usage{
CoyProcessor(data, name = "coy", id = "CaneID")
}
\arguments{
\item{data}{data frame containing the IDs and coy strings}

\item{name}{column containing the coy data, default = "coy"}

\item{id}{column containing the identifiers, default = "CaneID"}
}
\value{
a data frame
}
\description{
Returns a data frame with the standard COY analysis
}
